pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[project]
authors = [
  {name = "Mario Pascual González", email = "mpascual@uma.es"},
]
dependencies = [
  "numpy>=1.23.0",
  "numpy<2.0.0; sys_platform == 'darwin'", # macOS-specific constraint
  "matplotlib>=3.3.0",
  "opencv-python>=4.6.0",
  "pillow>=7.1.2",
  "pyyaml>=5.3.1",
  "requests>=2.23.0",
  "scipy>=1.4.1",
  "torch>=1.8.0",
  "torch>=1.8.0,!=2.4.0; sys_platform == 'win32'", # Windows-specific constraint
  "torchvision>=0.9.0",
  "tqdm>=4.64.0",
  "psutil",
  "py-cpuinfo",
  "pandas>=1.1.4",
  "seaborn>=0.11.0",
  "ultralytics-thop>=2.0.0",
  "optuna",
  "nvidia-ml-py3",
  "iterative-stratification",
  "natsort",
  "Scienceplots",
  "cmaes",
  "pandas",
  "types-requests",
  "types-PyYAML",
  "pycocotools",
  "requests",
  # shap
  # lightgbm
  # pyarrow
  # fastparquet
]
description = ""
name = "ICA_Detection"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1"

[project.urls]
Homepage = "https://github.com/MarioPasc/Coronary_Angiography_Detection"

[project.scripts] # Add this section
run_optimization = "ICA_Detection.optimization.utils.cli:main"
run_kfoldval_optimization = "ICA_Detection.optimization.engine.kfold_validation.engine:cli"


[tool.setuptools.packages.find]
include = ["ICA_Detection", "ICA_Detection.*", "ICA_Detection.external.ultralytics", "ICA_Detection.external.DCA_YOLOv8"]

[tool.setuptools.package-data]
"ICA_Detection.external.ultralytics" = ["*"]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "torchvision.*"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "pycocotools.*"
